moved child-specific methods to separate file
